Refactor Sewgit/Cash Bech32 (#64)
